requirement.txt
asgiref==3.8.1
dataclasses==0.6
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djongo==1.2.31
dnspython==2.7.0
drf-yasg==1.21.10
inflection==0.5.1
mongoengine==0.29.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pymongo==4.11.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
uritemplate==4.1.1
